backend/requirements.txt
fastapi
uvicorn
requests
transformers
torch
python-dotenv
openai